kaspresso params refactored
