+++ kaspresso/src/main/kotlin/com/kaspersky/kaspresso/autoscroll/AutoScrollProviderImpl.kt
@@ -4,6 +4,7 @@ import androidx.test.espresso.ViewInteraction
 import androidx.test.espresso.action.ViewActions
 import com.kaspersky.kaspresso.internal.extensions.other.isAllowed
 import com.kaspersky.kaspresso.logger.UiTestLogger
+import com.kaspersky.kaspresso.params.AutoScrollParams
 
 /**
  * The implementation of the [AutoScrollProvider] interface for [ViewInteraction]
