+++ kaspresso/src/main/kotlin/com/kaspersky/kaspresso/flakysafety/FlakySafetyProviderImpl.kt
@@ -3,6 +3,7 @@ package com.kaspersky.kaspresso.flakysafety
 import com.kaspersky.kaspresso.internal.extensions.other.isAllowed
 import com.kaspersky.kaspresso.internal.extensions.other.withMessage
 import com.kaspersky.kaspresso.logger.UiTestLogger
+import com.kaspersky.kaspresso.params.FlakySafetyParams
 import java.util.Timer
 import java.util.concurrent.locks.ReentrantLock
 import kotlin.concurrent.schedule
