+++ kaspresso/src/main/kotlin/com/kaspersky/kaspresso/interceptors/behavior/impl/autoscroll/AutoScrollWebBehaviorInterceptor.kt
@@ -1,7 +1,7 @@
 package com.kaspersky.kaspresso.interceptors.behavior.impl.autoscroll
 
 import androidx.test.espresso.web.sugar.Web
-import com.kaspersky.kaspresso.autoscroll.AutoScrollParams
+import com.kaspersky.kaspresso.params.AutoScrollParams
 import com.kaspersky.kaspresso.autoscroll.AutoScrollProvider
 import com.kaspersky.kaspresso.autoscroll.WebAutoScrollProviderImpl
 import com.kaspersky.kaspresso.interceptors.behavior.WebBehaviorInterceptor
