+++ kaspresso/src/main/kotlin/com/kaspersky/kaspresso/interceptors/behavior/impl/flakysafety/FlakySafeWebBehaviorInterceptor.kt
@@ -1,7 +1,7 @@
 package com.kaspersky.kaspresso.interceptors.behavior.impl.flakysafety
 
 import androidx.test.espresso.web.sugar.Web
-import com.kaspersky.kaspresso.flakysafety.FlakySafetyParams
+import com.kaspersky.kaspresso.params.FlakySafetyParams
 import com.kaspersky.kaspresso.flakysafety.FlakySafetyProvider
 import com.kaspersky.kaspresso.flakysafety.FlakySafetyProviderImpl
 import com.kaspersky.kaspresso.interceptors.behavior.WebBehaviorInterceptor
