+++ kaspresso/src/main/kotlin/com/kaspersky/kaspresso/kaspresso/Kaspresso.kt
@@ -6,7 +6,7 @@ import androidx.test.espresso.FailureHandler
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 import com.agoda.kakao.Kakao
-import com.kaspersky.kaspresso.autoscroll.AutoScrollParams
+import com.kaspersky.kaspresso.params.AutoScrollParams
 import com.kaspersky.kaspresso.device.Device
 import com.kaspersky.kaspresso.device.accessibility.Accessibility
 import com.kaspersky.kaspresso.device.accessibility.AccessibilityImpl
@@ -35,7 +35,7 @@ import com.kaspersky.kaspresso.device.screenshots.ScreenshotsImpl
 import com.kaspersky.kaspresso.device.server.AdbServer
 import com.kaspersky.kaspresso.device.server.AdbServerImpl
 import com.kaspersky.kaspresso.failure.LoggingFailureHandler
-import com.kaspersky.kaspresso.flakysafety.FlakySafetyParams
+import com.kaspersky.kaspresso.params.FlakySafetyParams
 import com.kaspersky.kaspresso.interceptors.behavior.DataBehaviorInterceptor
 import com.kaspersky.kaspresso.interceptors.behavior.ViewBehaviorInterceptor
 import com.kaspersky.kaspresso.interceptors.behavior.WebBehaviorInterceptor
@@ -70,6 +70,7 @@ import com.kaspersky.kaspresso.interceptors.watcher.view.impl.logging.LoggingVie
 import com.kaspersky.kaspresso.interceptors.watcher.view.impl.logging.LoggingWebAssertionWatcherInterceptor
 import com.kaspersky.kaspresso.logger.UiTestLogger
 import com.kaspersky.kaspresso.logger.UiTestLoggerImpl
+import com.kaspersky.kaspresso.params.Params
 import com.kaspersky.kaspresso.report.impl.AllureReportWriter
 
 /**
@@ -80,8 +81,7 @@ data class Kaspresso(
     internal val testLogger: UiTestLogger,
     internal val adbServer: AdbServer,
     internal val device: Device,
-    internal val flakySafetyParams: FlakySafetyParams,
-    internal val autoScrollParams: AutoScrollParams,
+    internal val params: Params,
     internal val viewActionWatcherInterceptors: List<ViewActionWatcherInterceptor>,
     internal val viewAssertionWatcherInterceptors: List<ViewAssertionWatcherInterceptor>,
     internal val atomWatcherInterceptors: List<AtomWatcherInterceptor>,
@@ -250,13 +250,15 @@ data class Kaspresso(
          * Holds the [FlakySafetyParams] for [com.kaspersky.kaspresso.flakysafety.FlakySafetyProvider]'s usage.
          * If it was not specified, the default implementation is used.
          */
-        var flakySafetyParams: FlakySafetyParams = FlakySafetyParams()
+        val flakySafetyParams: FlakySafetyParams =
+            FlakySafetyParams()
 
         /**
          * Holds the [AutoScrollParams] for [com.kaspersky.kaspresso.autoscroll.AutoScrollProvider]'s usage.
          * If it was not specified, the default implementation is used.
          */
-        var autoScrollParams: AutoScrollParams = AutoScrollParams()
+        val autoScrollParams: AutoScrollParams =
+            AutoScrollParams()
 
         /**
          * Holds the list of [ViewActionWatcherInterceptor]s.
@@ -405,8 +407,10 @@ data class Kaspresso(
                     exploit = exploit
                 ),
 
-                flakySafetyParams = flakySafetyParams,
-                autoScrollParams = autoScrollParams,
+                params = Params(
+                    flakySafetyParams = flakySafetyParams,
+                    autoScrollParams = autoScrollParams
+                ),
 
                 viewActionWatcherInterceptors = viewActionWatcherInterceptors,
                 viewAssertionWatcherInterceptors = viewAssertionWatcherInterceptors,
