+++ kaspresso/src/main/kotlin/com/kaspersky/kaspresso/params/Params.kt
@@ -1,0 +1,9 @@
+package com.kaspersky.kaspresso.params
+
+/**
+ * The facade class for all Kaspresso parameters.
+ */
+data class Params(
+    val flakySafetyParams: FlakySafetyParams,
+    val autoScrollParams: AutoScrollParams
+)
